requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.9
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-jazzmin==2.5.0
django-modeltranslation==0.18.5
django-oauth-toolkit==2.1.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
ecdsa==0.18.0
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
Interface==2.11.1
jwcrypto==1.3.1
Markdown==3.4.1
oauthlib==3.2.0
Pillow==9.2.0
protobuf==4.21.5
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pymemcache==3.5.2
pyparsing==3.0.9
python-dotenv==0.20.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tk==0.1.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
zope.schema==6.2.1
